fix: renamed files to avoid ambiguous imports, and fixed 'write' in _parquet.py
