+++ transparentpath/io/_pandas.py
