+++ transparentpath/io/_zipfile.py
